Device properties are now displayed and changes are written to server. Watch is used to poll for updates.
On start, a watch is created and href stored in localStorage. On next load, obelix verifies is the watch still exists and if it does, uses it instead of requesting new watch.
Some minor CSS fixes.
Added a new library URI.js for easier evaluation of relative hrefs.
Device box needs to move at least 5px for dragging to start.