.) Drop area for device boxes is now limited to viewport and device boxes can not be moved around to expand the viewport.
.) Redesign of jsPlumb endpoints and connections.

.) Finished tour.

Open points:

.) Sometimes connections cannot be restored when reloading the page.
.) Sometimes connections cannot be drawn (initialization of jsPlumb might be
the cause of the problem).
.) Sometimes connections cannot be deleted by double clicking on them.
.) Endpoints of a new connection are only rearranged when the source endpoint is located left to
the target endpoint.  If the source endpoint is located right to the target endpoint
one of the device boxes has to be moved so that the endpoints are rearranged.

changed obelix/source/index.haml
changed obelix/source/javascripts/app.js
changed obelix/source/stylesheets/app.scss