Bugfix: jsPlumb connections are not always drawn
An interval timer runs now (every 250 msec) executing
a function which checks if the required amount of endpoint
connection elements are visible in the browser. If so,
the interval timer is cancelled and the connections, if any,
are drawn.

The previous timer, which was responsible for calling jsPlumb
to draw the connections after an arbitrary amount of time (100 msec)
has been removed.

Additional edits:

Little code cleaning/formatting.

changed obelix/source/javascripts/app.js